Reorganised rs structure
