+++ examples/communication_child_to_parent/src/main.rs
@@ -5,10 +5,6 @@ use yew::{html, AttrValue, Callback, Component, Context, Html, Properties};
 mod child;
 mod parent;
 
-pub enum Msg {
-    ButtonClick(AttrValue),
-}
-
 fn main() {
     yew::Renderer::<Parent>::new().render();
 }
